tr1/game-flow: tidy the api
