+++ src/tr2/game/game_flow/sequencer.h
@@ -13,5 +13,3 @@ GAME_FLOW_COMMAND GF_DoDemoSequence(int32_t demo_num);
 GAME_FLOW_COMMAND GF_DoCutsceneSequence(int32_t cutscene_num);
 GAME_FLOW_COMMAND GF_DoLevelSequence(
     const GAME_FLOW_LEVEL *start_level, GAME_FLOW_SEQUENCE_CONTEXT seq_ctx);
-GAME_FLOW_COMMAND GF_RunLevel(
-    const GAME_FLOW_LEVEL *level, GAME_FLOW_SEQUENCE_CONTEXT seq_ctx);
