+++ src/tr2/meson.build
@@ -107,6 +107,7 @@ sources = [
   'game/game_flow/reader.c',
   'game/game_flow/inventory.c',
   'game/game_flow/sequencer.c',
+  'game/game_flow/sequencer_misc.c',
   'game/game_flow/vars.c',
   'game/gun/gun.c',
   'game/gun/gun_misc.c',
